+++ src/main/kotlin/com/unrec/imdb/loader/entity/BasicsEntity.kt
@@ -1,13 +1,13 @@
 package com.unrec.imdb.loader.entity
 
 data class BasicsEntity(
-    val titleId: Long,
+    val titleId: Int,
     val titleType: String,
     val primaryTitle: String,
     val originalTitle: String,
     val isAdult: Boolean?,
-    val startYear: Int?,
-    val endYear: Int?,
+    val startYear: Short?,
+    val endYear: Short?,
     val runtimeMinutes: Int?,
     val genres: String,
 ) : Entity()
