+++ src/main/kotlin/com/unrec/imdb/loader/entity/CrewEntity.kt
@@ -1,7 +1,7 @@
 package com.unrec.imdb.loader.entity
 
 data class CrewEntity(
-    val titleId: Long,
+    val titleId: Int,
     val directors: String?,
     val writers: String?,
 ) : Entity()
